Refactor various different aproaches to handling non-ascii strings from snmp This adds nav.Snmp.safestring() which is a mix of the existing decode_to_unicode and safestring functions
